Rename `nullMode.ts` to `nullTokenize.ts`
