+++ src/vs/editor/common/languages/nullTokenize.ts
