+++ src/vs/editor/common/languages/textToHtmlTokenizer.ts
@@ -7,7 +7,7 @@ import { CharCode } from 'vs/base/common/charCode';
 import * as strings from 'vs/base/common/strings';
 import { IViewLineTokens, LineTokens } from 'vs/editor/common/tokens/lineTokens';
 import { ILanguageIdCodec, IState, ITokenizationSupport, LanguageId, TokenizationRegistry } from 'vs/editor/common/languages';
-import { NullState, nullTokenizeEncoded } from 'vs/editor/common/languages/nullMode';
+import { NullState, nullTokenizeEncoded } from 'vs/editor/common/languages/nullTokenize';
 import { ILanguageService } from 'vs/editor/common/languages/language';
 
 export type IReducedTokenizationSupport = Omit<ITokenizationSupport, 'tokenize'>;
